+++ test/run_test.py
@@ -76,10 +76,7 @@ TESTS = [
     'distributed/rpc/faulty_agent/test_dist_autograd_spawn',
     'distributed/rpc/faulty_agent/test_rpc_spawn',
     'distributed/rpc/jit/test_dist_autograd_spawn',
-    'distributed/rpc/tensorpipe/test_dist_autograd_spawn',
-    'distributed/rpc/tensorpipe/test_dist_optimizer_spawn',
-    'distributed/rpc/tensorpipe/test_rpc_spawn',
-    'distributed/rpc/tensorpipe/test_ddp_under_dist_autograd',
+    'distributed/rpc/test_tensorpipe_agent',
     'distributed/rpc/test_dist_autograd_spawn',
     'distributed/rpc/test_dist_optimizer_spawn',
     'distributed/rpc/test_rpc_spawn',
@@ -98,10 +95,7 @@ WINDOWS_BLACKLIST = [
     'distributed/rpc/faulty_agent/test_rpc_spawn',
     'distributed/rpc/jit/test_dist_autograd_spawn',
     'distributed/rpc/jit/test_rpc_spawn',
-    'distributed/rpc/tensorpipe/test_dist_autograd_spawn',
-    'distributed/rpc/tensorpipe/test_dist_optimizer_spawn',
-    'distributed/rpc/tensorpipe/test_rpc_spawn',
-    'distributed/rpc/tensorpipe/test_ddp_under_dist_autograd',
+    'distributed/rpc/test_tensorpipe_agent',
     'distributed/rpc/test_dist_autograd_spawn',
     'distributed/rpc/test_dist_optimizer_spawn',
     'distributed/rpc/test_rpc_spawn',
@@ -116,10 +110,7 @@ ROCM_BLACKLIST = [
     'distributed/rpc/faulty_agent/test_rpc_spawn',
     'distributed/rpc/jit/test_dist_autograd_spawn',
     'distributed/rpc/jit/test_rpc_spawn',
-    'distributed/rpc/tensorpipe/test_dist_autograd_spawn',
-    'distributed/rpc/tensorpipe/test_dist_optimizer_spawn',
-    'distributed/rpc/tensorpipe/test_rpc_spawn',
-    'distributed/rpc/tensorpipe/test_ddp_under_dist_autograd',
+    'distributed/rpc/test_tensorpipe_agent',
     'distributed/rpc/test_dist_autograd_spawn',
     'distributed/test_ddp_under_dist_autograd',
     'distributed/rpc/test_dist_optimizer_spawn',
@@ -161,10 +152,7 @@ SLOW_TESTS = [
     'distributed/nn/jit/test_instantiator',
     'distributed/nn/api/test_remote_module_spawn',
     'distributed/test_distributed',
-    'distributed/rpc/tensorpipe/test_dist_autograd_spawn',
-    'distributed/rpc/tensorpipe/test_dist_optimizer_spawn',
-    'distributed/rpc/tensorpipe/test_rpc_spawn',
-    'distributed/rpc/tensorpipe/test_ddp_under_dist_autograd',
+    'distributed/rpc/test_tensorpipe_agent',
     'distributed/rpc/test_dist_autograd_spawn',
     'distributed/rpc/test_rpc_spawn',
     'distributed/test_ddp_under_dist_autograd',
